+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - `locale` and `posix-locale` for `--time-style=TIME_STYLE`.
 - Handle `mi` parameter of `$LS_COLORS`.
 - Lookup `TIME_STYLE` environment variable to determine the time format used.
+- `locale` and `clocale` for `--quoting-style=WORD`.
 
 ### Changed